Setup text completion functionality for convo files
This commit configures text completion for convo files. Setting up a match rule, turning off hitting enter indents, and assigning keyboard mappings to facilitate completion via the GPTComplete() function. It also creates the SetupForConvo() function, which is called when a convo buffer is entered. It adds the $put . <cr> command, which adds a new line when ffj is pressed.