Add Server and SSH configuration for Clipboard forwarding
Adds a setup for clipboard forwarding from the SSH server to localhost via a RemoteForward configuration to running the `ncat` command on the localhost. A pbcopy server  is setup on port 2224, and a pbpaste server is setup on port 2225. The `rpbcopy` and `rpbpaste` files are also added here to cleanly send/receive data to/from the clipboard on the remote server. Finally, the `ssh_config` file was updated with a `lightsail_clipboard` entry with the RemoteForward configurations.